refactor(event): move handle `channel_deleted` to channelsEventsService
